Detect SEOPress Pro, and pull all metas
This keeps non-pro users functional, because the `proschemas` and `proschemasManual` only exist if SEOPress Pro is detected. Furthermore, SEOPress Pro is detected by the existence of the `seopress_pro_license_status` option, which only returns valid when both the Pro plugin in installed, and the license is validated. Ref: #7

This also pulls all metas listed on the SEOpress site. Ref: #2